[REFACTOR] #3 디자인 시스템 component preview를 별도의 파일로 분리
