+++ core/ui/src/main/java/pokitmons/pokit/core/ui/components/block/category_card/CategoryCard.kt
@@ -2,11 +2,9 @@ package pokitmons.pokit.core.ui.components.block.category_card
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
@@ -22,7 +20,6 @@ import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import pokitmons.pokit.core.ui.R
 import pokitmons.pokit.core.ui.components.block.category_card.attributes.CategoryCardState
@@ -123,25 +120,6 @@ private fun getSubTextColor(
         }
         CategoryCardState.DISABLE -> {
             PokitTheme.colors.textDisable
-        }
-    }
-}
-
-
-
-@Preview(showBackground = true)
-@Composable
-fun CategoryCardPreview() {
-    PokitTheme {
-        Column(
-            modifier = Modifier.fillMaxSize(),
-            verticalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            CategoryCard(state = CategoryCardState.DISABLE, item = "STRING", painter = painterResource(id = R.drawable.icon_24_star), title = "카테고리입니당", sub = "15개 항목", onClickKebab = {}, onClickItem = {})
-
-            CategoryCard(state = CategoryCardState.DEFAULT, item = "STRING", painter = painterResource(id = R.drawable.icon_24_star), title = "카테고리입니당", sub = "15개 항목", onClickKebab = {}, onClickItem = {})
-
-            CategoryCard(state = CategoryCardState.ACTIVE, item = "STRING", painter = painterResource(id = R.drawable.icon_24_star), title = "카테고리입니당", sub = "15개 항목", onClickKebab = {}, onClickItem = {})
         }
     }
 }
